[Worker Registration Step 3: PR Status and CI Failure Workers](2a) Make worker registry generic
