+++ packages/execution-engine/src/__tests__/worker-registry.test.ts
@@ -1,15 +1,17 @@
 import { describe, it, expect } from 'vitest';
 
-import { RECOVERY_WORKER_KIND, type AutoFixRecoveryStore, type AutoFixRecoverySubmitter } from '../auto-fix-recovery.js';
 import {
   AUTO_FIX_WORKER_KIND,
-  CI_FAILURE_WORKER_KIND,
-  createWorkerRegistry,
-  PR_STATUS_WORKER_KIND,
+  RECOVERY_WORKER_KIND,
   registerAutoFixWorker,
-  registerBuiltinWorkers,
-  type WorkerRuntimeDependencies,
-} from '../worker-registry.js';
+  type AutoFixRecoveryStore,
+  type AutoFixRecoverySubmitter,
+} from '../auto-fix-recovery.js';
+import { registerBuiltinWorkers } from '../builtin-workers.js';
+import type { WorkerRuntimeDependencies } from '../worker-runtime-dependencies.js';
+import { createWorkerRegistry } from '../worker-registry.js';
+import { CI_FAILURE_WORKER_KIND } from '../workers/ci-failure-worker.js';
+import { PR_STATUS_WORKER_KIND } from '../workers/pr-status-worker.js';
 
 const silentLogger = {
   debug: () => {},
@@ -35,13 +37,13 @@ function deps(): WorkerRuntimeDependencies {
 
 describe('worker registry', () => {
   it('starts empty before anything is registered', () => {
-    const registry = createWorkerRegistry();
+    const registry = createWorkerRegistry<WorkerRuntimeDependencies>();
     expect(registry.list()).toEqual([]);
     expect(registry.get(AUTO_FIX_WORKER_KIND)).toBeUndefined();
   });
 
   it('returns the auto-fix definition by its kind once registered', () => {
-    const registry = registerAutoFixWorker(createWorkerRegistry());
+    const registry = registerAutoFixWorker(createWorkerRegistry<WorkerRuntimeDependencies>());
 
     const definition = registry.get(AUTO_FIX_WORKER_KIND);
     expect(definition).toBeDefined();
@@ -53,7 +55,7 @@ describe('worker registry', () => {
   });
 
   it('registers every built-in worker in one call', () => {
-    const registry = registerBuiltinWorkers(createWorkerRegistry());
+    const registry = registerBuiltinWorkers(createWorkerRegistry<WorkerRuntimeDependencies>());
 
     expect(registry.list().map((d) => d.kind)).toEqual([
       AUTO_FIX_WORKER_KIND,
@@ -65,12 +67,12 @@ describe('worker registry', () => {
     expect(registry.get(CI_FAILURE_WORKER_KIND)).toBeDefined();
   });
   it('returns nothing for an unknown kind', () => {
-    const registry = registerAutoFixWorker(createWorkerRegistry());
+    const registry = registerAutoFixWorker(createWorkerRegistry<WorkerRuntimeDependencies>());
     expect(registry.get('does-not-exist')).toBeUndefined();
   });
 
   it('builds a recovery worker runtime from the auto-fix factory', () => {
-    const registry = registerAutoFixWorker(createWorkerRegistry());
+    const registry = registerAutoFixWorker(createWorkerRegistry<WorkerRuntimeDependencies>());
     const definition = registry.get(AUTO_FIX_WORKER_KIND);
     expect(definition).toBeDefined();
 
@@ -83,7 +85,7 @@ describe('worker registry', () => {
 
 
   it('builds the PR status and CI-failure worker runtimes from the registered factories', () => {
-    const registry = registerBuiltinWorkers(createWorkerRegistry());
+    const registry = registerBuiltinWorkers(createWorkerRegistry<WorkerRuntimeDependencies>());
 
     expect(registry.get(PR_STATUS_WORKER_KIND)?.factory(deps()).identity.kind).toBe(PR_STATUS_WORKER_KIND);
     expect(registry.get(CI_FAILURE_WORKER_KIND)?.factory(deps()).identity.kind).toBe(CI_FAILURE_WORKER_KIND);
